feat: validators for users entity
