+++ src/config/routes.js
@@ -2,7 +2,7 @@ const { Router } = require('express');
 const healthcheck = require('../routes/healthcheck');
 const restaurants = require('../routes/restaurants');
 const withList = require('../routes/wishList');
-const users = require('../routes/user');
+const users = require('../routes/users');
 const auth = require('../middlewares/auth');
 
 module.exports = (app) => {
