+++ src/routes/restaurants.js
@@ -1,13 +1,13 @@
 const restaurant = require('../controllers/restaurants');
-const { restaurantValidations } = require('../config/validations');
+const { restaurantValidator } = require('../config/validator');
 
 module.exports = (router) => {
   router
-    .post('/restaurants', restaurantValidations, restaurant.save)
+    .post('/restaurants', restaurantValidator, restaurant.save)
     .get('/restaurants', restaurant.findAll);
 
   router
-    .put('/restaurants/:id', restaurantValidations, restaurant.update)
+    .put('/restaurants/:id', restaurantValidator, restaurant.update)
     .get('/restaurants/:id', restaurant.findById)
     .delete('/restaurants/:id', restaurant.remove);
 };
