--- src/routes/user.js
@@ -1,9 +1,0 @@
-const users = require('../controllers/users');
-
-module.exports = (router) => {
-  router
-    .post('/signin', users.save);
-
-  router
-    .post('/auth', users.auth);
-};
