+++ src/routes/wishList.js
@@ -1,13 +1,13 @@
 const wishList = require('../controllers/wishList');
-const { wishValidations } = require('../config/validations');
+const { wishValidator } = require('../config/validator');
 
 module.exports = (router) => {
   router
     .get('/wishlist', wishList.findAll)
-    .post('/wishlist', wishValidations, wishList.save);
+    .post('/wishlist', wishValidator, wishList.save);
 
   router
-    .put('/wishlist/:id', wishValidations, wishList.update)
+    .put('/wishlist/:id', wishValidator, wishList.update)
     .delete('/wishlist/:id', wishList.remove)
     .get('/wishlist/:id', wishList.findById);
 };
